+++ tests/_data/postgres_post.sql
@@ -378,6 +378,28 @@ INSERT INTO "oauth2_scope" VALUES
         true,
         EXTRACT(EPOCH FROM NOW()),
         EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1005019,
+        'applied-by-default-if-requested',
+        'applied if requested by client without user confirmation',
+        null,
+        3,
+        false,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1005020,
+        'applied-by-default-if-requested-for-client',
+        'applied if requested by client without user confirmation (not applied itself, but applied in oauth2_client_scope)',
+        null,
+        0,
+        false,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
     );
 ;
 
@@ -517,6 +539,24 @@ INSERT INTO "oauth2_client_scope" VALUES
         EXTRACT(EPOCH FROM NOW()),
         EXTRACT(EPOCH FROM NOW())
     ),
+    (
+        1003000, -- 'test-client-type-auth-code-valid'
+        1005019, -- 'applied-by-default-if-requested'
+        null,
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003000, -- 'test-client-type-auth-code-valid'
+        1005020, -- 'applied-by-default-if-requested-for-client'
+        3,
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
     -- test-client-type-client-credentials-valid
     (
         1003001, -- 'test-client-type-client-credentials-valid'
@@ -613,6 +653,24 @@ INSERT INTO "oauth2_client_scope" VALUES
         1003002, -- 'test-client-type-password-public-valid'
         1005018, -- 'not-required-has-been-rejected-before'
         null,
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003002, -- 'test-client-type-password-public-valid'
+        1005019, -- 'applied-by-default-if-requested'
+        null,
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003002, -- 'test-client-type-password-public-valid'
+        1005020, -- 'applied-by-default-if-requested-for-client'
+        3,
         null,
         true,
         EXTRACT(EPOCH FROM NOW()),
@@ -723,6 +781,60 @@ INSERT INTO "oauth2_client_scope" VALUES
         1003007, -- 'test-client-type-auth-code-open-id-connect-skip-authorization'
         current_setting('Yii2Oauth2Server.PostMigration.oidcOfflineAccessScopeId')::int,
         0, -- Note: for testing the offline access scope is not auto applied
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005000, -- 'user.id.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005001, -- 'user.username.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005002, -- 'user.email_address.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005003, -- 'user.enabled.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005004, -- 'user.created_at.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
+        null,
+        true,
+        EXTRACT(EPOCH FROM NOW()),
+        EXTRACT(EPOCH FROM NOW())
+    ),
+    (
+        1003008, -- 'test-client-type-personal-access-token'
+        1005005, -- 'user.updated_at.read'
+        3, -- APPLIED_BY_DEFAULT_IF_REQUESTED
         null,
         true,
         EXTRACT(EPOCH FROM NOW()),